Work around Android bugs.
Includes:
  * Check environ against NULL before using it, since Android sets it to
    NULL.
  * Cast st_size of struct stat to off_t. Android uses long long for
    st_size as its off_t is only 32 bit, but st_size should be off_t
    according to POSIX.
  * Android's strtod() does not accept 0x, as specified by C99. Thus,
    the test is disabled if __ANDROID__ is defined.

+++ src/OFFile.m
@@ -407,7 +407,8 @@ parse_mode(const char *mode)
 		@throw [OFOpenFileFailedException exceptionWithPath: path
 							       mode: @"r"];
 
-	return s.st_size;
+	/* On Android, off_t is 32 bit, but st_size is long long there */
+	return (off_t)s.st_size;
 }
 
 + (OFDate*)modificationDateOfFileAtPath: (OFString*)path
